requirements.txt
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.7
inflect == 2.1.0
manage.py==0.2.10
mysqlclient==1.3.13
psycopg2==2.7.6.1
pytz==2018.6
requests==2.20.0
urllib3==1.24.1
whitenoise==4.1.1
djangorestframework==3.9.0